updated app

+++ FitnessApp.Api/Models/Exercise.cs
@@ -8,6 +8,6 @@
 
         public string Name { get; set; }
 
-        public int Order { get; set; }
+        public ushort Order { get; set; }
     }
 }
